+++ main.go
@@ -32,15 +32,19 @@ func main() {
 		upstream              string
 		label                 string
 		enableLabelAPIs       bool
+		nonAPIPathPassthrough bool
 	)
 
 	flagset := flag.NewFlagSet(os.Args[0], flag.ExitOnError)
 	flagset.StringVar(&insecureListenAddress, "insecure-listen-address", "", "The address the prom-label-proxy HTTP server should listen on.")
 	flagset.StringVar(&upstream, "upstream", "", "The upstream URL to proxy to.")
-	flagset.StringVar(&label, "label", "", "The label to enforce in all proxied PromQL queries.")
+	flagset.StringVar(&label, "label", "", "The label to enforce in all proxied PromQL queries. This label will be also required as the URL parameter to get the value to be injected."+
+		"For example: -label=tenant will make it required for this proxy to have URL in form of: <URL>?tenant=abc&other_params...")
 	flagset.BoolVar(&enableLabelAPIs, "enable-label-apis", false, "When specified proxy allows to inject label to label APIs like /api/v1/labels and /api/v1/label/<name>/values."+
 		"NOTE: Enable with care. Selection of matcher is still in development, see https://github.com/thanos-io/thanos/issues/3351 and https://github.com/prometheus/prometheus/issues/6178. If enabled and"+
 		"any labels endpoint does not support selectors, injected matcher will be silently dropped.")
+	flagset.BoolVar(&nonAPIPathPassthrough, "non-api-path-passthrough", false, "If enabled prom-label-proxy will proxy all requests to non recognized paths (except /api/*) to the upstream, without any enforcing."+
+		"Use carefully as it can easily cause a data leak if upstream have other read APIs that prom-label-proxy is not aware.")
 
 	//nolint: errcheck // Parse() will exit on error.
 	flagset.Parse(os.Args[1:])
@@ -57,7 +61,15 @@ func main() {
 		log.Fatalf("Invalid scheme for upstream URL %q, only 'http' and 'https' are supported", upstream)
 	}
 
-	routes := injectproxy.NewRoutes(upstreamURL, label, injectproxy.WithEnabledLabelsAPI())
+	var opts []injectproxy.Option
+	if enableLabelAPIs {
+		opts = append(opts, injectproxy.WithEnabledLabelsAPI())
+	}
+	if nonAPIPathPassthrough {
+		opts = append(opts, injectproxy.WithNonAPIPassThrough())
+	}
+	routes := injectproxy.NewRoutes(upstreamURL, label, opts...)
+
 	mux := http.NewServeMux()
 	mux.Handle("/", routes)