+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 * [FEATURE] Add support  /api/v1/series, /api/v1/labels and /api/v1/label/<name>/values (Prometheus/Thanos).
+* [FEATURE] Add -non-api-path-passthrough flag (disabled by default), which allows
+ exposing all non API resources from upstream without enforcing (e.g Prometheus UI).
 
 ## 0.2.0 / 2020-10-08